backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
httpx
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
google-generativeai
apscheduler
yt-dlp
pydantic